requirements.txt
# Backend dependencies
Flask
Flask-Cors
Flask-JWT-Extended
Flask-MySQL
Flask-PyMongo
openai
requests
pandas
scikit-learn
nltk
spacy
gunicorn
matplotlib
seaborn
python-dotenv
bcrypt

# Frontend dependencies
streamlit
